+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project uses [Semantic Versioning](http://semver.org/).
 
+
+# [3.0.0] - 2020-07-23
+# Changed
+ - renamed repo to stork
+
+========= prior development under the name apparate =========
+
 # [2.3.0] - 2020-07-22
 # Added
  - repo rename warning